[MOVE] 패턴 기간 enum 경로 이동 및 연관 수정 #3
