+++ src/main/java/com/synergyx/trading/enums/PeriodUnit.java
@@ -1,4 +1,4 @@
-package com.synergyx.trading.model;
+package com.synergyx.trading.enums;
 
 public enum PeriodUnit {
     SEC, MINUTE, HOUR, DAY, MONTH;
